gun/pistols: move Gun_Pistols_Draw to libtrx
