+++ src/tr1/game/gun/gun_pistols.h
@@ -6,7 +6,7 @@
 
 #include "global/types.h"
 
-void Gun_Pistols_Draw(LARA_GUN_TYPE weapon_type);
-void Gun_Pistols_Ready(LARA_GUN_TYPE weapon_type);
+#include <libtrx/game/gun/pistols.h>
+
 void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_Animate(LARA_GUN_TYPE weapon_type);
